+++ src/main/java/com/virtusa/interview/converter/util/Utility.java
@@ -2,7 +2,7 @@ package com.virtusa.interview.converter.util;
 
 import org.springframework.stereotype.Component;
 
-import com.virtusa.interview.converter.contant.Constants;
+import com.virtusa.interview.converter.constant.Constants;
 
 
 /**
@@ -18,22 +18,22 @@ public class Utility {
    	 * @param number the number
    	 * @return the string
    	 */
-   	public String convertLessThanOneThousand(Integer number) {
+   	public String convertLessThanOneThousand(long number) {
 	        String current;
 	        
 	        if (number % 100 < 20){
-	            current = Constants.NUM_NAMES[number % 100];
+	            current = Constants.NUM_NAMES[(int) (number % 100)];
 	            number /= 100;
 	        }
 	        else {
-	            current = Constants.NUM_NAMES[number % 10];
+	            current = Constants.NUM_NAMES[(int) (number % 10)];
 	            number /= 10;
 	            
-	            current = Constants.TENS_NAMES[number % 10] + current;
+	            current = Constants.TENS_NAMES[(int) (number % 10)] + current;
 	            number /= 10;
 	        }
 	        if (number == 0) return current;
-	        return Constants.NUM_NAMES[number] + " hundred" + current;
+	        return Constants.NUM_NAMES[(int) number] + " hundred" + current;
 	    }
 
 }
